requirements.txt
numpy==1.21.6
rdkit-pypi==2022.3.5
h5py==2.10.0
matplotlib==3.2.2
tensorflow==2.3.0